detect invalid translations
